api-server/requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
blinker==1.4
click==7.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.3.0
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
priority==1.3.0
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
pytest-asyncio==0.14.0
python-dotenv==0.15.0
Quart==0.14.1
regex==2020.11.13
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wsproto==1.0.0
